Scroll to bottom of chat on open
