+++ client/src/components/Chat/MessageCreator.js
@@ -2,7 +2,6 @@ import React from 'react';
 import {
     Box,
     Button,
-    Grid,
     Container,
     IconButton,
     TextField
